Refactor location of CrashTestDataProvider
